refactor: 추천 관련 dto들을 별도 패키지로 분리
