+++ src/main/java/umc/cockple/demo/domain/exercise/dto/recommendation/ExerciseRecommendationCalendarDTO.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.exercise.dto;
+package umc.cockple.demo.domain.exercise.dto.recommendation;
 
 import lombok.Builder;
 import umc.cockple.demo.domain.exercise.enums.MyPartyExerciseOrderType;
