+++ src/main/java/umc/cockple/demo/domain/exercise/repository/ExerciseRepositoryCustom.java
@@ -1,7 +1,7 @@
 package umc.cockple.demo.domain.exercise.repository;
 
 import umc.cockple.demo.domain.exercise.domain.Exercise;
-import umc.cockple.demo.domain.exercise.dto.ExerciseRecommendationCalendarDTO;
+import umc.cockple.demo.domain.exercise.dto.recommendation.ExerciseRecommendationCalendarDTO;
 
 import java.time.LocalDate;
 import java.util.List;
